Python dictionaries are not ordered
The dictionary grouping can be out of order since python dictionaries
are not guarateed to be always ordered the same. This causes dash
dict states to be mangled in some environments.

Here we sort the dictionaries in both the indice calculation and the
grouping calculation to make them ordered correctly.

Although dictionary ordering is guaranteed from python 3.7 onwards
https://mail.python.org/pipermail/python-dev/2017-December/151283.html
I'm seeing issues in some overly optimized environments.